requirements.txt
fqdn==1.5.1
gdown==5.2.0
isoduration==20.11.0
langchain-experimental==0.3.4
langchain-openai==0.3.33
langchain-community>=0.3.0
matplotlib==3.10.6
networkx>=3.0
notebook==7.4.5
plotly==6.3.0
PySocks==1.7.1
python-dotenv>=1.0.0
rfc3987-syntax==1.1.0
streamlit==1.50.0
tinycss2==1.4.0
Unidecode==1.4.0
uri-template==1.3.0
watchdog==6.0.0
webcolors==24.11.1
